fix(packet): Grammar whitespace
Move populate into parser
