+++ packages/parser/src/language/packet/packet.langium
@@ -15,6 +15,5 @@ PacketBlock:
     start=INT('-' end=INT)? ':' label=STRING NEWLINE+
 ;
 
-hidden terminal WS: /\s+/;
 terminal INT returns number: /0|[1-9][0-9]*/;
 terminal STRING: /"[^"]*"|'[^']*'/;
